requirements.txt
pandas
requests
ipython-sql
sqlalchemy
psycopg2 
python-dotenv
ydata_profiling
matplotlib>=3.2, <=3.7.3
great_expectations
dbt-postgres